Extract SigmaParser and deps into a new module
